+++ src/EnrolmentWorkflow/Request/class.UsersAssistantsAjaxAutoCompleteCtrl.php
@@ -1,0 +1,55 @@
+<?php
+
+namespace srag\Plugins\SrUserEnrolment\EnrolmentWorkflow\Request;
+
+use ilSrUserEnrolmentPlugin;
+use srag\CustomInputGUIs\SrUserEnrolment\MultiSelectSearchNewInputGUI\UsersAjaxAutoCompleteCtrl;
+use srag\Plugins\SrUserEnrolment\Utils\SrUserEnrolmentTrait;
+
+/**
+ * Class UsersAssistantsAjaxAutoCompleteCtrl
+ *
+ * @package srag\Plugins\SrUserEnrolment\EnrolmentWorkflow\Request
+ *
+ * @author  studer + raimann ag - Team Custom 1 <support-custom1@studer-raimann.ch>
+ */
+class UsersAssistantsAjaxAutoCompleteCtrl extends UsersAjaxAutoCompleteCtrl
+{
+
+    use SrUserEnrolmentTrait;
+    const PLUGIN_CLASS_NAME = ilSrUserEnrolmentPlugin::class;
+    /**
+     * @var RequestsGUI
+     */
+    protected $parent;
+
+
+    /**
+     * UsersAssistantsAjaxAutoCompleteCtrl constructor
+     *
+     * @param RequestsGUI $parent
+     */
+    public function __construct(RequestsGUI $parent)
+    {
+        parent::__construct();
+
+        $this->parent = $parent;
+    }
+
+
+    /**
+     * @inheritDoc
+     */
+    protected function formatUsers(array $users) : array
+    {
+        return parent::formatUsers(array_filter($users, function (array $user) : bool {
+            return self::srUserEnrolment()
+                ->enrolmentWorkflow()
+                ->requests()
+                ->canRequestWithAssistant($this->parent->getObjRefId(), current(self::srUserEnrolment()->enrolmentWorkflow()->steps()->getSteps(self::srUserEnrolment()
+                    ->enrolmentWorkflow()
+                    ->selectedWorkflows()
+                    ->getWorkflowId(self::dic()->objDataCache()->lookupObjId($this->parent->getObjRefId()))))->getStepId(), self::dic()->user()->getId(), $user["usr_id"]);
+        }));
+    }
+}